login

+++ doit/src/main/java/it/unicam/ids/doit/controller/EspertoController.java
@@ -4,6 +4,7 @@ import com.sun.istack.NotNull;
 import it.unicam.ids.doit.entity.Progetto;
 import it.unicam.ids.doit.service.EspertoService;
 import it.unicam.ids.doit.service.ProgettoService;
+import it.unicam.ids.doit.service.UserHandlerService;
 import it.unicam.ids.doit.service.impl.EspertoServiceImpl;
 import it.unicam.ids.doit.service.impl.ProgettoServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,18 +16,20 @@ import java.util.List;
 public class EspertoController {
 
     @Autowired
-    private ProgettoService progettoService;
-    @Autowired
     private EspertoService espertoService;
+    @Autowired
+    private UserHandlerService userHandlerService;
 
     /**
      * metodo che ricerca tutti i progetti che sono i stato di Waiting
      * @return lista dei progetti da valutare
      */
+    @CrossOrigin(origins = "http://localhost:4200")
     @GetMapping(value="/getprogettidavalutare/")
     @ResponseBody
-    public List<Progetto> getprogettidavalutare(/*@RequestParam Long idEsperto*/){
-        return progettoService.getAllProgettiValutare();
+    public List<Progetto> getprogettidavalutare(@RequestParam @NotNull Long idEsperto, @RequestParam @NotNull Long token ){
+        if(userHandlerService.check(idEsperto,token))   return espertoService.getAllProgettiValutare();
+        return null;
     }
 
     /**
@@ -34,10 +37,12 @@ public class EspertoController {
      * @param id id del progetto
      * @return progetto con l'id, ritorna tutti i dati
      */
+    @CrossOrigin(origins = "http://localhost:4200")
     @GetMapping(value="/getprogettidavalutare/{id}")
     @ResponseBody
-    public Progetto getprogetto(@PathVariable Long id){
-        return progettoService.getProgetto(id);
+    public Progetto getprogetto(@PathVariable Long id, @RequestParam @NotNull Long idEsperto , @RequestParam @NotNull Long token ){
+        if(userHandlerService.check(idEsperto,token))  return espertoService.getProgetto(id);
+        return null;
     }
 
     /**
@@ -46,12 +51,16 @@ public class EspertoController {
      * @param idEsperto id di colui che ha accettato
      * @return errore oppure messaggio di successo
      */
+    @CrossOrigin(origins = "http://localhost:4200")
     @PostMapping(value="/getprogettidavalutare/{idProgetto}/confirm")
     @ResponseBody
-    public String confirmProgetto(@PathVariable Long idProgetto, @RequestParam @NotNull Long idEsperto){
+    public String confirmProgetto(@PathVariable Long idProgetto, @RequestParam @NotNull Long idEsperto,@RequestParam @NotNull Long token){
         try {
-            espertoService.confirmProgetto(idProgetto,idEsperto);
-            return "success";
+            if(userHandlerService.check(idEsperto,token)) {
+                espertoService.confirmProgetto(idProgetto, idEsperto);
+                return "success";
+            }
+            return "not logged";
         }catch(Exception e){
             return e.getMessage();
         }
@@ -63,12 +72,16 @@ public class EspertoController {
      * @param idEsperto id di colui che ha rifiutato
      * @return errore oppure messaggio di successo
      */
+    @CrossOrigin(origins = "http://localhost:4200")
     @PostMapping(value="/getprogettidavalutare/{idProgetto}/decline")
     @ResponseBody
-    public String declineProgetto(@PathVariable Long idProgetto, @RequestParam @NotNull Long idEsperto){
+    public String declineProgetto(@PathVariable Long idProgetto, @RequestParam @NotNull Long idEsperto,@RequestParam @NotNull Long token){
         try{
-            espertoService.declineProgetto(idProgetto, idEsperto); //TODO se mettiamo confirm e decline dentro al service dell'esperto?
-            return "success";
+            if(userHandlerService.check(idEsperto,token)) {
+                espertoService.declineProgetto(idProgetto, idEsperto); //TODO se mettiamo confirm e decline dentro al service dell'esperto?
+                return "success";
+            }
+            return "not logged";
         }catch(Exception e){
             return e.getMessage();
         }
@@ -79,12 +92,16 @@ public class EspertoController {
      * @param idEsperto id del esperto da eliminare
      * @return messaggio di successo o insuccesso
      */
+    @CrossOrigin(origins = "http://localhost:4200")
     @DeleteMapping(value="/remove")
     @ResponseBody
-    public String removeAccount(@RequestParam @NotNull Long idEsperto){
+    public String removeAccount(@RequestParam @NotNull Long idEsperto,@RequestParam @NotNull Long token){
         try{
-            espertoService.deleteEsperto(idEsperto);
-            return "Eliminato correttamente";
+            if(userHandlerService.check(idEsperto,token)) {
+                espertoService.deleteEsperto(idEsperto);
+                return "Eliminato correttamente";
+            }
+            return "Not logged";
         }catch (Exception e){
             return e.getMessage();
         }
